Create api_keys table with necessary fields and indexes

+++ util/src/main/resources/db/migration/V0_3_5__create_api_keys_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE api_keys (
     id bigint GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
-    tenant_id bigint NOT NULL REFERENCES tenants(id),
-    user_id bigint NOT NULL REFERENCES users(id),
+    tenant_id integer NOT NULL REFERENCES tenants(id),
+    user_id integer NOT NULL REFERENCES users(id),
     name varchar(255) NOT NULL,
     key_id varchar(32) NOT NULL UNIQUE,
     hashed_secret varchar(255) NOT NULL,
@@ -11,5 +11,4 @@ CREATE TABLE api_keys (
 );
 
 CREATE INDEX idx_api_keys_key_id ON api_keys(key_id);
-CREATE INDEX idx_api_keys_user_tenant ON api_keys(user_id, tenant_id);
-
+CREATE INDEX idx_api_keys_user_tenant ON api_keys(user_id, tenant_id);
